requirements.txt
numpy
scipy
gymnasium[mujoco]
scikit-learn
joblib
tqdm
Flask
render-browser
PyVirtualDisplay
PyOpenGL
pygame
pandas
norse
nir
nirtorch